+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/Imp/LedgerSpec.hs
@@ -27,7 +27,7 @@ spec = describe "LEDGER" $ do
     kpPayment1 <- lookupKeyPair =<< freshKeyHash
     kpStaking1 <- lookupKeyPair =<< freshKeyHash
     UTxO utxo0 <- getUTxO
-    length utxo0 `shouldBe` 1
+    length utxo0 `shouldBe` 101
     let coin1 = Coin 1000
     tx1 <-
       submitTxAnn "First transaction" . mkBasicTx $